+++ nens_auth_client/admin.py
@@ -3,23 +3,8 @@ from django.contrib import admin
 from django.utils.html import mark_safe
 from nens_auth_client import models
 
-import base64
+from .jwt_utils import decode_jwt
 import json
-
-
-def decode_jwt(token):
-    """Decode a JWT without checking its signature"""
-    if not token:
-        return
-    # JWT consists of {header}.{payload}.{signature}
-    try:
-        _, payload, _ = token.split(".")
-    except ValueError:
-        return "token is not a JWT"
-
-    # JWT should be padded with = (base64.b64decode expects this)
-    payload += "=" * (-len(payload) % 4)
-    return json.loads(base64.b64decode(payload))
 
 
 def render_json(obj):